requirements.txt
ffmpeg==1.4
ffmpeg_python==0.2.0
Flask==3.1.2
openai==2.2.0
python-dotenv==1.1.1
